Refactor: move `UserService` to deployment subsystem.
