+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/domain/NotifyUserService.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.studies.domain
+package dk.cachet.carp.deployment.domain
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
