+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/domain/users/UserRepository.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.studies.domain.users
+package dk.cachet.carp.deployment.domain.users
 
 import dk.cachet.carp.common.UUID
