+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/domain/users/Username.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.studies.domain.users
+package dk.cachet.carp.deployment.domain.users
 
 import kotlinx.serialization.Serializable
