+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/infrastructure/UserServiceRequest.kt
@@ -1,14 +1,14 @@
-package dk.cachet.carp.studies.infrastructure
+package dk.cachet.carp.deployment.infrastructure
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.common.ddd.createServiceInvoker
 import dk.cachet.carp.common.ddd.createServiceInvokerOverloaded
 import dk.cachet.carp.common.ddd.ServiceInvoker
-import dk.cachet.carp.studies.application.UserService
-import dk.cachet.carp.studies.domain.users.Account
-import dk.cachet.carp.studies.domain.users.Participant
-import dk.cachet.carp.studies.domain.users.Username
+import dk.cachet.carp.deployment.application.UserService
+import dk.cachet.carp.deployment.domain.users.Account
+import dk.cachet.carp.deployment.domain.users.Participant
+import dk.cachet.carp.deployment.domain.users.Username
 import kotlinx.serialization.Serializable
