+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/domain/NotifyUserServiceMock.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.studies.domain
+package dk.cachet.carp.deployment.domain
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
