+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/infrastructure/AccountTest.kt
@@ -1,8 +1,8 @@
-package dk.cachet.carp.studies.infrastructure
+package dk.cachet.carp.deployment.infrastructure
 
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.studies.domain.users.Account
-import dk.cachet.carp.studies.domain.users.Participant
+import dk.cachet.carp.deployment.domain.users.Account
+import dk.cachet.carp.deployment.domain.users.Participant
 import kotlin.test.*
