+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/infrastructure/UserServiceRequestsTest.kt
@@ -1,11 +1,11 @@
-package dk.cachet.carp.studies.infrastructure
+package dk.cachet.carp.deployment.infrastructure
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.common.ddd.ServiceInvoker
-import dk.cachet.carp.studies.application.UserService
-import dk.cachet.carp.studies.application.UserServiceMock
-import dk.cachet.carp.studies.domain.users.Username
+import dk.cachet.carp.deployment.application.UserService
+import dk.cachet.carp.deployment.application.UserServiceMock
+import dk.cachet.carp.deployment.domain.users.Username
 import dk.cachet.carp.test.runBlockingTest
 import kotlin.test.*
