+++ carp.studies.core/src/commonMain/kotlin/dk/cachet/carp/studies/infrastructure/Serialization.kt
@@ -3,33 +3,17 @@
 package dk.cachet.carp.studies.infrastructure
 
 import dk.cachet.carp.common.serialization.createDefaultJSON
+import dk.cachet.carp.deployment.infrastructure.DEPLOYMENT_SERIAL_MODULE
+import dk.cachet.carp.protocols.infrastructure.PROTOCOLS_SERIAL_MODULE
 import dk.cachet.carp.studies.domain.StudyDescription
 import dk.cachet.carp.studies.domain.StudyOwner
 import dk.cachet.carp.studies.domain.StudySnapshot
 import dk.cachet.carp.studies.domain.StudyStatus
-import dk.cachet.carp.studies.domain.users.Account
-import dk.cachet.carp.studies.domain.users.AccountIdentity
-import dk.cachet.carp.studies.domain.users.EmailAccountIdentity
-import dk.cachet.carp.studies.domain.users.Participant
-import dk.cachet.carp.studies.domain.users.Username
-import dk.cachet.carp.studies.domain.users.UsernameAccountIdentity
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.modules.EmptyModule
 import kotlinx.serialization.modules.plus
-import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.modules.SerialModule
 
-
-/**
- * Types in the [dk.cachet.carp.studies] module which need to be registered when using [Json] serializer.
- */
-val STUDIES_SERIAL_MODULE = SerializersModule {
-    polymorphic( AccountIdentity::class )
-    {
-        UsernameAccountIdentity::class with UsernameAccountIdentity.serializer()
-        EmailAccountIdentity::class with EmailAccountIdentity.serializer()
-    }
-}
 
 /**
  * Create a [Json] serializer adopting a default CARP infrastructure configuration with all [dk.cachet.carp.studies] types registered.
@@ -38,7 +22,7 @@ val STUDIES_SERIAL_MODULE = SerializersModule {
  */
 fun createStudiesSerializer( module: SerialModule = EmptyModule ): Json
 {
-    return createDefaultJSON( STUDIES_SERIAL_MODULE + module )
+    return createDefaultJSON( PROTOCOLS_SERIAL_MODULE + DEPLOYMENT_SERIAL_MODULE + module )
 }
 
 /**
@@ -47,42 +31,6 @@ fun createStudiesSerializer( module: SerialModule = EmptyModule ): Json
  * [createStudiesSerializer] can be used to this end, by including all extending types in the [SerialModule] as parameter.
  */
 var JSON: Json = createStudiesSerializer()
-
-/**
- * Create a [Account] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
- */
-fun Account.Companion.fromJson( json: String ): Account =
-    JSON.parse( serializer(), json )
-
-/**
- * Serialize to JSON, using the globally set infrastructure serializer ([JSON]).
- */
-fun Account.toJson(): String =
-    JSON.stringify( Account.serializer(), this )
-
-/**
- * Create a [Username] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
- */
-fun Username.Companion.fromJson( json: String ): Username =
-    JSON.parse( serializer(), json )
-
-/**
- * Serialize to JSON, using the globally set infrastructure serializer ([JSON]).
- */
-fun Username.toJson(): String =
-    JSON.stringify( Username.serializer(), this )
-
-/**
- * Create a [Participant] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
- */
-fun Participant.Companion.fromJson( json: String ): Participant =
-    JSON.parse( serializer(), json )
-
-/**
- * Serialize to JSON, using the globally set infrastructure serializer ([JSON]).
- */
-fun Participant.toJson(): String =
-    JSON.stringify( Participant.serializer(), this )
 
 /**
  * Create a [StudySnapshot] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
